+++ CHANGELOG.md
@@ -9,11 +9,11 @@ Bug fixes:
 
 Compatibility:
 * Implement `String#append_as_bytes` for Ruby 3.4 (#4087, @Earlopain).
-
 * Update error message of `Array#fetch` with key out of range (#4085, @herwinw).
 * Implement `Array#fetch_values` for Ruby 3.4 (#4085, @herwinw).
 * Remove deprecated `Refinement#refined_class` method for Ruby 3.4 (#4088, @herwinw).
 * Update `Hash#inspect` for Ruby 3.4 (#4090, @Earlopain).
+* Implement `Fiber.[]`, `Fiber.[]=`, and `Fiber#{storage,storage=}` (#4007, @nirvdrum).
 
 Performance: